admin/requirements.txt
streamlit==1.41.0
firebase-admin==6.6.0
plotly==5.24.0
python-dotenv==1.0.1
pandas==2.2.3
Pillow==11.1.0
